backend/python/requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.3.0

# Database
psycopg2-binary==2.9.7
pymongo==4.5.0
motor==3.3.1

# MQTT and IoT
paho-mqtt==1.6.1
asyncio-mqtt==0.13.0

# BIM Processing
# ifcopenshell==0.6.0  # Will be installed separately via system packages
numpy==1.24.3
scipy==1.11.1
trimesh==3.20.0

# Background Tasks
celery==5.3.1
redis==4.6.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Web Scraping and HTTP
requests==2.31.0
aiohttp==3.8.5

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1 